♻️ refactor(openapi): reorganize openapi integration code
- move client configuration to separate HttpxClientConfig class
- update all references from ClientConfig to HttpxClientConfig
- organize imports in __init__.py
- simplify integration.py by removing redundant code
- fix import path for OpenAPIIntegration in tool_registry.py
