+++ src/toolregistry/openapi/__init__.py
@@ -1,0 +1,12 @@
+from .integration import OpenAPIIntegration, OpenAPITool, OpenAPIToolWrapper
+from .utils import HttpxClientConfig, load_openapi_spec, load_openapi_spec_async
+
+__all__ = [
+    "OpenAPIIntegration",
+    "OpenAPITool",
+    "OpenAPIToolWrapper",
+    # Helpers and utilities for OpenAPI integration.
+    "HttpxClientConfig",
+    "load_openapi_spec",
+    "load_openapi_spec_async",
+]
